+++ backend/src/main/java/org/rdfarchitect/rdf/graph/wrapper/GraphRewindableWithUUIDs.java
@@ -18,18 +18,10 @@
 package org.rdfarchitect.rdf.graph.wrapper;
 
 import org.apache.jena.graph.Graph;
-import org.apache.jena.graph.Node;
 import org.apache.jena.query.ReadWrite;
-import org.apache.jena.rdf.model.Model;
-import org.apache.jena.rdf.model.ModelFactory;
-import org.apache.jena.rdf.model.Resource;
-import org.apache.jena.rdf.model.Statement;
-import org.apache.jena.vocabulary.RDFS;
 import org.jetbrains.annotations.NotNull;
 import org.rdfarchitect.exception.graph.GraphNotInATransactionException;
 import org.rdfarchitect.exception.graph.GraphTransactionException;
-import org.rdfarchitect.models.cim.rdf.resources.CIMS;
-import org.rdfarchitect.models.cim.rdf.resources.RDFA;
 import org.rdfarchitect.rdf.graph.DeltaCompressible;
 import org.rdfarchitect.rdf.graph.GraphUtils;
 
@@ -70,94 +62,5 @@ public class GraphRewindableWithUUIDs extends GraphRewindable {
             compressBase();
         }
         logger.debug("Committed transaction.");
-    }
-
-    public static void removeUUIDs(Graph graph) {
-        graph.find(Node.ANY, RDFA.uuid.asNode(), Node.ANY).toList().forEach(graph::delete);
-    }
-
-    public static void correctPackagePrefix(Graph graph, boolean usePackagePrefix) {
-        var model = ModelFactory.createModelForGraph(graph);
-        var packages = model.listResourcesWithProperty(RDF.type, CIMS.classCategory).toList();
-
-        final var PREFIX = "Package_";
-
-        for (var packageResource : packages) {
-            var labelStmt = packageResource.getProperty(RDFS.label);
-            if (labelStmt == null) {
-                continue;
-            }
-
-            var currentLabel = labelStmt.getString();
-            var currentUri = packageResource.getURI();
-            var sepIdx = Math.max(currentUri.lastIndexOf('#'), currentUri.lastIndexOf('/'));
-            var uriBase = currentUri.substring(0, sepIdx + 1);
-            var uriSuffix = currentUri.substring(sepIdx + 1);
-
-            String newLabel;
-            String newUriSuffix;
-
-            if (usePackagePrefix) {
-                newLabel = currentLabel.startsWith(PREFIX) ? currentLabel : PREFIX + currentLabel;
-                newUriSuffix = uriSuffix.startsWith(PREFIX) ? uriSuffix : PREFIX + uriSuffix;
-            } else {
-                newLabel =
-                        currentLabel.startsWith(PREFIX)
-                                ? currentLabel.substring(PREFIX.length())
-                                : currentLabel;
-                newUriSuffix =
-                        uriSuffix.startsWith(PREFIX)
-                                ? uriSuffix.substring(PREFIX.length())
-                                : uriSuffix;
-            }
-
-            var newUri = uriBase + newUriSuffix;
-            var uriChanged = !newUri.equals(currentUri);
-            var labelChanged = !newLabel.equals(currentLabel);
-
-            if (uriChanged) {
-                changeURIAndLabel(packageResource, labelChanged, newUri, newLabel, model);
-            } else if (labelChanged) {
-                changeLabel(packageResource, labelStmt, newLabel);
-            }
-        }
-    }
-
-    private static void changeURIAndLabel(
-            Resource packageResource,
-            boolean labelChanged,
-            String newUri,
-            String newLabel,
-            Model model) {
-        var newResource = model.createResource(newUri);
-        packageResource
-                .listProperties()
-                .toList()
-                .forEach(
-                        stmt -> {
-                            if (stmt.getPredicate().equals(RDFS.label) && labelChanged) {
-                                changeLabel(newResource, stmt, newLabel);
-                            } else {
-                                newResource.addProperty(stmt.getPredicate(), stmt.getObject());
-                            }
-                        });
-
-        model.listStatements(null, null, packageResource)
-                .toList()
-                .forEach(stmt -> model.add(stmt.getSubject(), stmt.getPredicate(), newResource));
-        model.listStatements(null, null, packageResource).toList().forEach(model::remove);
-
-        model.removeAll(packageResource, null, null);
-    }
-
-    private static void changeLabel(
-            Resource packageResource, Statement labelStmt, String newLabel) {
-        var lang = labelStmt.getLanguage();
-        packageResource.removeAll(RDFS.label);
-        if (lang != null && !lang.isEmpty()) {
-            packageResource.addProperty(RDFS.label, newLabel, lang);
-        } else {
-            packageResource.addProperty(RDFS.label, newLabel);
-        }
     }
 }
